Create AT1G01100.fasta
If you run "01_get_data.py" this file will be generated automaticly whci contains cds sequences of orthologue genes but not nessasirily unique sequences so first you will need to run "01_make_sequence_titles_unique.py". on the other hand your gene of interest sequence is not in this file. to add that your gene of interest so to add your gene of interest to this fasta file you will need to run "02_combine_cds_and_gene_of_interest.py"